requirements.txt
asgiref==3.9.1
Django==5.2.5
django-nested-admin==4.1.2
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.2
